% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeCount.R
\name{NormalizeCount}
\alias{NormalizeCount}
\title{Remove Unwanted Variation from RNA-Seq Data}
\usage{
NormalizeCount(x, HKGs, k)
}
\arguments{
\item{x}{Expression matrix}

\item{HKGs}{Housekeeping genes indices}

\item{k}{Number of factors to remove}
}
\value{
A list with normalized counts and factors
}
\description{
Remove Unwanted Variation from RNA-Seq Data
}
